backend/requirements.txt
astroid==2.3.3
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.4.4
pymongo==3.10.1
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
google-cloud-storage==1.26.0